feat: Optimize solving process with timer and logging
Refactor the solving process in the main function to optimize performance. Replace the usage of the time module with the ttimer library to measure the execution time of the optimizer.solve() method. Add logging using the loguru library to display the timer results. This improves the efficiency and provides better visibility into the execution time of the solving process.